+++ docs/release_notes.rst
@@ -14,6 +14,7 @@ Release notes for **Shopit**.
 * Re-work the reviews, making them non-translatable. Not compatible with the old reviews, make sure to save them
   (if you have any) before upgrading. A way for adding reviews was not provided before so this should not be the case.
 * Add setting `SHOPIT_REVIEW_ACTIVE_DEFAULT`. This decides if created reviews are active by default.
+* Handle updating shopping cart via ajax, add success messages to it.
 
 .. attention::